+++ frontend/features/CreateTaskConfirm/constants.ts
@@ -1,0 +1,7 @@
+export const priorityMap: Record<string, string> = {
+    very_low: "とても低い",
+    low: "低い",
+    normal: "普通",
+    high: "高い",
+    very_high: "とても高い",
+  };
